requirements.txt
telethon>=1.36,<2.0
APScheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
ollama>=0.6.1,<1.0
